campus-doc-agent/requirements.txt
openai==0.28.1
python-docx==0.8.11
streamlit==1.29.0
python-dotenv==1.0.0